Smash it all together
