--- templates/10-calico.conf
@@ -1,18 +1,0 @@
-{
-    "name": "calico-k8s-network",
-    "type": "calico",
-    "etcd_endpoints": "{{ connection_string }}",
-    "etcd_key_file": "{{ etcd_key_path }}",
-    "etcd_cert_file": "{{ etcd_cert_path }}",
-    "etcd_ca_cert_file": "{{ etcd_ca_path }}",
-    "log_level": "info",
-    "ipam": {
-        "type": "calico-ipam"
-    },
-    "policy": {
-        "type": "k8s"
-    },
-    "kubernetes": {
-        "kubeconfig": "{{ kubeconfig_path }}"
-    }
-}
