+++ templates/10-canal.conf
@@ -1,0 +1,18 @@
+{
+    "name": "cdk-canal",
+    "type": "flannel",
+    "delegate": {
+        "type": "calico",
+        "etcd_endpoints": "{{ connection_string }}",
+        "etcd_key_file": "{{ etcd_key_path }}",
+        "etcd_cert_file": "{{ etcd_cert_path }}",
+        "etcd_ca_cert_file": "{{ etcd_ca_path }}",
+        "log_level": "info",
+        "policy": {
+            "type": "k8s"
+        },
+        "kubernetes": {
+            "kubeconfig": "{{ kubeconfig_path }}"
+        }
+    }
+}
