--- templates/10-flannel.conf
@@ -1,7 +1,0 @@
-{
-    "name": "CDK-flannel-network",
-    "type": "flannel",
-    "delegate": {
-      "isDefaultGateway": true
-    }
-}
