+++ templates/calico-node.service
@@ -18,7 +18,7 @@ ExecStart=/usr/bin/docker run --net=host --privileged --name=calico-node \
   -e AS= \
   -e CALICO_LIBNETWORK_ENABLED=true \
   -e IP6= \
-  -e CALICO_NETWORKING_BACKEND=bird \
+  -e CALICO_NETWORKING_BACKEND=none \
   -e FELIX_DEFAULTENDPOINTTOHOSTACTION=ACCEPT \
   -v /var/run/calico:/var/run/calico \
   -v /lib/modules:/lib/modules \
