--- templates/pool.yaml
@@ -1,9 +1,0 @@
-apiVersion: v1
-kind: ipPool
-metadata:
-  cidr: {{ cidr }}
-spec:
-  ipip:
-    enabled: {{ ipip }}
-    mode: always
-  nat-outgoing: {{ nat_outgoing }}
